// frontend/src/api-client/models/PlaylistGameCount.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Reelshelf | v1
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface PlaylistGameCount
 */
export interface PlaylistGameCount {
  /**
   *
   * @type {string}
   * @memberof PlaylistGameCount
   */
  gameCategoryId: string;
  /**
   *
   * @type {number}
   * @memberof PlaylistGameCount
   */
  clipCount: number;
}

/**
 * Check if a given object implements the PlaylistGameCount interface.
 */
export function instanceOfPlaylistGameCount(
  value: object,
): value is PlaylistGameCount {
  if (!("gameCategoryId" in value) || value["gameCategoryId"] === undefined)
    return false;
  if (!("clipCount" in value) || value["clipCount"] === undefined) return false;
  return true;
}

export function PlaylistGameCountFromJSON(json: any): PlaylistGameCount {
  return PlaylistGameCountFromJSONTyped(json, false);
}

export function PlaylistGameCountFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): PlaylistGameCount {
  if (json == null) {
    return json;
  }
  return {
    gameCategoryId: json["game_category_id"],
    clipCount: json["clip_count"],
  };
}

export function PlaylistGameCountToJSON(json: any): PlaylistGameCount {
  return PlaylistGameCountToJSONTyped(json, false);
}

export function PlaylistGameCountToJSONTyped(
  value?: PlaylistGameCount | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    game_category_id: value["gameCategoryId"],
    clip_count: value["clipCount"],
  };
}